// src/app/Interfaces/injury.ts
export interface Injury {
    id?: number;
    date: string;
    day: string;
    time: string;
    location: string;
    memberName: string;
    membershipNo: string;
    injuryType: string;
    cause: string;
    action: string;
    safetyOfficer: string;
    medic: string;
    control: string;
    supervisor: string;
    notes: string;
}